requirements.txt
requests
python-dotenv
pydub
tqdm
SpeechRecognition
google-cloud-speech
Pillow
